+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/CredentialPublicKeys.java
@@ -1,6 +1,6 @@
 package com.concordium.sdk.transactions;
 
-import com.concordium.sdk.responses.accountinfo.credential.Key;
+import com.concordium.sdk.crypto.ed25519.ED25519PublicKey;
 import lombok.Getter;
 import lombok.SneakyThrows;
 import lombok.ToString;
@@ -8,30 +8,40 @@ import lombok.val;
 import java.nio.ByteBuffer;
 import java.util.Map;
 
+/**
+ * The public credential keys belonging to a credential holder
+ */
 @Getter
 @ToString
 public class CredentialPublicKeys {
-    private final Map<Index, Key> keys;
+    private final Map<Index, ED25519PublicKey> keys;
     private final int threshold;
 
-    CredentialPublicKeys(Map<Index, Key> keys,
+    CredentialPublicKeys(Map<Index, ED25519PublicKey> keys,
                          int threshold) {
         this.keys = keys;
         this.threshold = threshold;
     }
 
-    public static CredentialPublicKeys from(Map<Index, Key> keys, int threshold) {
+    public static CredentialPublicKeys from(Map<Index, ED25519PublicKey> keys, int threshold) {
         return new CredentialPublicKeys(keys, threshold);
+    }
+
+    byte[] getSchemeIdBytes() {
+        val buffer = ByteBuffer.allocate(PublicKey.BYTES);
+        buffer.put(PublicKey.ED25519.getValue());
+        return buffer.array();
     }
 
     @SneakyThrows
     public byte[] getBytes() {
         val keysLenBytes = keys.keySet().size();
         int keyBufferSize = TransactionType.BYTES;
+        val schemeIdBytes = getSchemeIdBytes();
         for (Index key : keys.keySet())
         {
-            Key value = keys.get(key);
-            keyBufferSize += TransactionType.BYTES + value.getBytes().length;
+            ED25519PublicKey value = keys.get(key);
+            keyBufferSize += TransactionType.BYTES + value.getBytes().length + PublicKey.BYTES;
         }
         val buffer = ByteBuffer.allocate(
                         keyBufferSize +
@@ -39,7 +49,8 @@ public class CredentialPublicKeys {
         buffer.put((byte)keysLenBytes);
         for (Index key : keys.keySet())
         {
-            Key value = keys.get(key);
+            ED25519PublicKey value = keys.get(key);
+            buffer.put(schemeIdBytes);
             buffer.put(key.getValue());
             buffer.put(value.getBytes());
         }
