+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/PublicKey.java
@@ -2,6 +2,10 @@ package com.concordium.sdk.transactions;
 
 import lombok.Getter;
 
+/**
+ * Public verification key for a given signature scheme.
+ */
+
 public enum PublicKey {
     ED25519((byte) 0);
